requirements.txt
PyQt5
RPi.GPIO
hcsr04sensor
pyserial